requirements.txt
requests
csv
json
openslide
hashlib
PyYAML
